Break out services, finish node service
Add stubs for missing streams. NodeService should be complete now. (May
still need to remove some unnecessary data items.)

Break out remaining services. Begin putting in place data model for
services.
